# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Highleads API",
        "description": "API highleads-backend",
        "contact": {
            "name": "Marco Aurélio",
            "email": "marcoascabral@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/notes/{contato}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Retorna as notas inseridas para um Contato",
                "description": "Retorna as notas inseridas para um Contato",
                "operationId": "18dd74f16ff2d1670ccbd9532b123716",
                "parameters": [
                    {
                        "name": "contato",
                        "in": "path",
                        "description": "contato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados."
                    },
                    "404": {
                        "description": "Item não encontrado."
                    }
                }
            }
        },
        "/v1/notes": {
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Criar um novo registro",
                "description": "Criar um novo registro de Notas",
                "operationId": "eee9f395e1de7e73f46daec89c233afb",
                "requestBody": {
                    "description": "Dados da Nota",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contatos_fk",
                                    "usuarios_fk",
                                    "notas"
                                ],
                                "properties": {
                                    "contatos_fk": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "usuarios_fk": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "notas": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "orgi urbi sanctum distro"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2199999999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nota salva com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "created": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao inserir."
                    }
                }
            }
        },
        "/v1/notes/{id}": {
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Exclui uma nota",
                "description": "Exclui uma nota por seu ID.",
                "operationId": "50d791ddf97a8701c20ab1b3e3ed9b67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Nota excluída com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "excluído": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nota não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}